requirements.txt
langchain
python-dotenv
langchain-community
# langchain-groq
sentence_transformers
# pypdf
# pymupdf
bs4
# arxiv
# wikipedia
# langchain-text-splitters
langchain-ollama
# langchain-openai
# chromadb
langchain-huggingface
# faiss-cpu
# langchain-chroma
# duckdb
pandas
# openai
streamlit
torch 
transformers
# unstructured
# unstructured[pdf]